release

+++ services/charts/lib/controller/SetupController.class.php
@@ -7,21 +7,29 @@ class controller_SetupController extends ufront_web_mvc_Controller {
 		$this->mongo = $mongo;
 	}}
 	public $mongo;
-	public function dropRenderables() {
+	public function dropRenderables($auth) {
+		$this->authorize($auth);
 		$this->dropCollection("renderables");
-		return $this->redirectToStatus();
+		return $this->redirectToStatus($auth);
 	}
-	public function dropCache() {
+	public function authorize($auth) {
+		if($auth !== "6kdsbgv46272") {
+			throw new HException(new ufront_web_error_UnauthorizedError(_hx_anonymous(array("fileName" => "SetupController.hx", "lineNumber" => 30, "className" => "controller.SetupController", "methodName" => "authorize"))));
+		}
+	}
+	public function dropCache($auth) {
+		$this->authorize($auth);
 		$this->dropCollection("cache");
-		return $this->redirectToStatus();
+		return $this->redirectToStatus($auth);
 	}
-	public function dropCollections() {
+	public function dropCollections($auth) {
+		$this->authorize($auth);
 		$this->dropCollection("renderables");
 		$this->dropCollection("cache");
-		return $this->redirectToStatus();
+		return $this->redirectToStatus($auth);
 	}
-	public function redirectToStatus() {
-		return new ufront_web_mvc_ForwardResult(null, _hx_anonymous(array("controller" => "setup", "action" => "mongodb")));
+	public function redirectToStatus($auth) {
+		return new ufront_web_mvc_ForwardResult(null, _hx_anonymous(array("controller" => "setup", "action" => "mongodb", "auth" => $auth)));
 	}
 	public function dropCollection($collection) {
 		$db = new mongo_MongoDB($this->mongo->m->selectDB("chartsrenderer1")); $collection1 = new mongo_MongoCollection($db->db->selectCollection($collection));
@@ -35,8 +43,9 @@ class controller_SetupController extends ufront_web_mvc_Controller {
 		$dbname = "chartsrenderer1"; $db = new mongo_MongoDB($this->mongo->m->selectDB($dbname));
 		return new mongo_MongoCollection($db->db->selectCollection("renderables"));
 	}
-	public function createCollections() {
-		$dbname = "chartsrenderer1"; $db = new mongo_MongoDB($this->mongo->m->selectDB($dbname)); $cacheCollections = new _hx_array($db->db->listCollections()); $renderablesCollectionName = "renderables"; $cacheCollectionName = "cache";
+	public function createCollections($auth) {
+		$this->authorize($auth);
+		$dbname = "chartsrenderer1"; $db = new mongo_MongoDB($this->mongo->m->selectDB($dbname)); $cacheCollections = new _hx_array($db->db->listCollections()); $renderablesCollectionName = "renderables"; $cacheCollectionName = "cache"; $configCollectionName = "config";
 		$renderableCollection = new mongo_MongoCollection($db->db->selectCollection($renderablesCollectionName));
 		if(php_Lib::objectOfAssociativeArray($renderableCollection->c->validate())->ok < 1) {
 			$renderableCollection = new mongo_MongoCollection($db->db->createCollection($renderablesCollectionName));
@@ -46,12 +55,22 @@ class controller_SetupController extends ufront_web_mvc_Controller {
 		$cacheCollection = new mongo_MongoCollection($db->db->selectCollection($cacheCollectionName));
 		if(php_Lib::objectOfAssociativeArray($cacheCollection->c->validate())->ok < 1) {
 			$cacheCollection = new mongo_MongoCollection($db->db->createCollection($cacheCollectionName));
-			$renderableCollection->ensureIndexOn("uid", _hx_anonymous(array("unique" => true)));
+			$cacheCollection->ensureIndexOn("uid", _hx_anonymous(array("unique" => true)));
 			$cacheCollection->ensureIndexOn("expiresOn", null);
 		}
-		return new ufront_web_mvc_ForwardResult(null, _hx_anonymous(array("controller" => "setup", "action" => "mongodb")));
+		$configCollection = new mongo_MongoCollection($db->db->selectCollection($configCollectionName));
+		if(php_Lib::objectOfAssociativeArray($cacheCollection->c->validate())->ok < 1) {
+			$configCollection = new mongo_MongoCollection($db->db->createCollection($configCollectionName));
+			$configCollection->ensureIndexOn("name", _hx_anonymous(array("unique" => true)));
+		}
+		$controller = new controller_RenderableAPIController(new model_RenderableGateway($renderableCollection));
+		$renderable = $controller->makeRenderable(model_Sample::$html, model_Sample::$config);
+		$config = new model_ConfigGateway($configCollection);
+		$config->setSampleUID($renderable->getUid());
+		return $this->redirectToStatus($auth);
 	}
-	public function mongodb() {
+	public function mongodb($auth) {
+		$this->authorize($auth);
 		$dbname = "chartsrenderer1"; $db = new mongo_MongoDB($this->mongo->m->selectDB($dbname)); $cacheCollections = new _hx_array($db->db->listCollections()); $renderablesCollectionName = "renderables"; $cacheCollectionName = "cache"; $renderablesExists = true; $cacheExists = true;
 		$renderableCollection = new mongo_MongoCollection($db->db->selectCollection($renderablesCollectionName));
 		if(php_Lib::objectOfAssociativeArray($renderableCollection->c->validate())->ok < 1) {
@@ -64,35 +83,41 @@ class controller_SetupController extends ufront_web_mvc_Controller {
 		$content = _hx_anonymous(array("baseurl" => App::baseUrl(), "url" => new ufront_web_mvc_view_UrlHelperInst($this->controllerContext->requestContext), "db" => _hx_anonymous(array("name" => $dbname, "collections" => $cacheCollections)), "renderables" => _hx_anonymous(array("name" => $renderablesCollectionName, "exists" => $renderablesExists, "count" => (($renderablesExists) ? $renderableCollection->c->count() : -1))), "cache" => _hx_anonymous(array("name" => $cacheCollectionName, "exists" => $cacheExists, "count" => (($cacheExists) ? $cacheCollection->c->count() : -1)))));
 		return new ufront_web_mvc_ContentResult(_hx_deref(new template_MongoDBStatus())->execute($content), null);
 	}
-	public function topRenderables($top) {
+	public function topRenderables($auth, $top) {
 		if($top === null) {
 			$top = 10;
 		}
+		$this->authorize($auth);
 		$gate = new model_RenderableGateway($this->renderableCollection()); $list = $gate->topByUsage($top); $content = _hx_anonymous(array("baseurl" => App::baseUrl(), "url" => new ufront_web_mvc_view_UrlHelperInst($this->controllerContext->requestContext), "top" => $top, "renderables" => $list));
 		return new ufront_web_mvc_ContentResult(_hx_deref(new template_RenderablesInfo())->execute($content), null);
 	}
-	public function purge() {
+	public function purge($auth) {
+		$this->authorize($auth);
 		$gate = new model_CacheGateway($this->cacheCollection());
 		$gate->removeExpired();
 		$gate1 = new model_RenderableGateway($this->renderableCollection());
 		$gate1->removeOldAndUnused(null);
-		return $this->redirectToStatus();
+		return $this->redirectToStatus($auth);
 	}
-	public function purgeCache() {
+	public function purgeCache($auth) {
+		$this->authorize($auth);
 		$gate = new model_CacheGateway($this->cacheCollection()); $purged = $gate->removeExpired();
-		return $this->redirectToStatus();
+		return $this->redirectToStatus($auth);
 	}
-	public function clearCache() {
+	public function clearCache($auth) {
+		$this->authorize($auth);
 		$gate = new model_CacheGateway($this->cacheCollection()); $purged = $gate->removeAll();
-		return $this->redirectToStatus();
+		return $this->redirectToStatus($auth);
 	}
-	public function purgeRenderables() {
+	public function purgeRenderables($auth) {
+		$this->authorize($auth);
 		$gate = new model_RenderableGateway($this->renderableCollection()); $purged = $gate->removeOldAndUnused(null);
-		return $this->redirectToStatus();
+		return $this->redirectToStatus($auth);
 	}
-	public function purgeExpiredRenderables() {
+	public function purgeExpiredRenderables($auth) {
+		$this->authorize($auth);
 		$gate = new model_RenderableGateway($this->renderableCollection()); $purged = $gate->removeExpired();
-		return $this->redirectToStatus();
+		return $this->redirectToStatus($auth);
 	}
 	public function info() {
 		return $this->collectPhpInfo();
@@ -112,6 +137,6 @@ class controller_SetupController extends ufront_web_mvc_Controller {
 		else
 			throw new HException('Unable to call «'.$m.'»');
 	}
-	static $__rtti = "<class path=\"controller.SetupController\" params=\"\">\x0A\x09<extends path=\"ufront.web.mvc.Controller\"/>\x0A\x09<mongo><c path=\"mongo.Mongo\"/></mongo>\x0A\x09<dropRenderables public=\"1\" set=\"method\" line=\"20\"><f a=\"\"><c path=\"ufront.web.mvc.ForwardResult\"/></f></dropRenderables>\x0A\x09<dropCache public=\"1\" set=\"method\" line=\"26\"><f a=\"\"><c path=\"ufront.web.mvc.ForwardResult\"/></f></dropCache>\x0A\x09<dropCollections public=\"1\" set=\"method\" line=\"32\"><f a=\"\"><c path=\"ufront.web.mvc.ForwardResult\"/></f></dropCollections>\x0A\x09<redirectToStatus set=\"method\" line=\"39\"><f a=\"\"><c path=\"ufront.web.mvc.ForwardResult\"/></f></redirectToStatus>\x0A\x09<dropCollection set=\"method\" line=\"44\"><f a=\"collection\">\x0A\x09<c path=\"String\"/>\x0A\x09<e path=\"Void\"/>\x0A</f></dropCollection>\x0A\x09<cacheCollection set=\"method\" line=\"51\"><f a=\"\"><c path=\"mongo.MongoCollection\"/></f></cacheCollection>\x0A\x09<renderableCollection set=\"method\" line=\"58\"><f a=\"\"><c path=\"mongo.MongoCollection\"/></f></renderableCollection>\x0A\x09<createCollections public=\"1\" set=\"method\" line=\"65\"><f a=\"\"><c path=\"ufront.web.mvc.ForwardResult\"/></f></createCollections>\x0A\x09<mongodb public=\"1\" set=\"method\" line=\"95\"><f a=\"\"><c path=\"ufront.web.mvc.ContentResult\"/></f></mongodb>\x0A\x09<topRenderables public=\"1\" set=\"method\" line=\"143\"><f a=\"?top\">\x0A\x09<c path=\"Int\"/>\x0A\x09<c path=\"ufront.web.mvc.ContentResult\"/>\x0A</f></topRenderables>\x0A\x09<purge public=\"1\" set=\"method\" line=\"156\"><f a=\"\"><c path=\"ufront.web.mvc.ForwardResult\"/></f></purge>\x0A\x09<purgeCache public=\"1\" set=\"method\" line=\"165\"><f a=\"\"><c path=\"ufront.web.mvc.ForwardResult\"/></f></purgeCache>\x0A\x09<clearCache public=\"1\" set=\"method\" line=\"172\"><f a=\"\"><c path=\"ufront.web.mvc.ForwardResult\"/></f></clearCache>\x0A\x09<purgeRenderables public=\"1\" set=\"method\" line=\"179\"><f a=\"\"><c path=\"ufront.web.mvc.ForwardResult\"/></f></purgeRenderables>\x0A\x09<purgeExpiredRenderables public=\"1\" set=\"method\" line=\"186\"><f a=\"\"><c path=\"ufront.web.mvc.ForwardResult\"/></f></purgeExpiredRenderables>\x0A\x09<info public=\"1\" set=\"method\" line=\"193\"><f a=\"\"><c path=\"String\"/></f></info>\x0A\x09<collectPhpInfo set=\"method\" line=\"198\"><f a=\"\"><c path=\"String\"/></f></collectPhpInfo>\x0A\x09<new public=\"1\" set=\"method\" line=\"14\"><f a=\"mongo\">\x0A\x09<c path=\"mongo.Mongo\"/>\x0A\x09<e path=\"Void\"/>\x0A</f></new>\x0A</class>";
+	static $__rtti = "<class path=\"controller.SetupController\" params=\"\">\x0A\x09<extends path=\"ufront.web.mvc.Controller\"/>\x0A\x09<mongo><c path=\"mongo.Mongo\"/></mongo>\x0A\x09<dropRenderables public=\"1\" set=\"method\" line=\"20\"><f a=\"auth\">\x0A\x09<c path=\"String\"/>\x0A\x09<c path=\"ufront.web.mvc.ForwardResult\"/>\x0A</f></dropRenderables>\x0A\x09<authorize set=\"method\" line=\"27\"><f a=\"auth\">\x0A\x09<c path=\"String\"/>\x0A\x09<e path=\"Void\"/>\x0A</f></authorize>\x0A\x09<dropCache public=\"1\" set=\"method\" line=\"33\"><f a=\"auth\">\x0A\x09<c path=\"String\"/>\x0A\x09<c path=\"ufront.web.mvc.ForwardResult\"/>\x0A</f></dropCache>\x0A\x09<dropCollections public=\"1\" set=\"method\" line=\"40\"><f a=\"auth\">\x0A\x09<c path=\"String\"/>\x0A\x09<c path=\"ufront.web.mvc.ForwardResult\"/>\x0A</f></dropCollections>\x0A\x09<redirectToStatus set=\"method\" line=\"48\"><f a=\"auth\">\x0A\x09<c path=\"String\"/>\x0A\x09<c path=\"ufront.web.mvc.ForwardResult\"/>\x0A</f></redirectToStatus>\x0A\x09<dropCollection set=\"method\" line=\"53\"><f a=\"collection\">\x0A\x09<c path=\"String\"/>\x0A\x09<e path=\"Void\"/>\x0A</f></dropCollection>\x0A\x09<cacheCollection set=\"method\" line=\"60\"><f a=\"\"><c path=\"mongo.MongoCollection\"/></f></cacheCollection>\x0A\x09<renderableCollection set=\"method\" line=\"67\"><f a=\"\"><c path=\"mongo.MongoCollection\"/></f></renderableCollection>\x0A\x09<createCollections public=\"1\" set=\"method\" line=\"74\"><f a=\"auth\">\x0A\x09<c path=\"String\"/>\x0A\x09<c path=\"ufront.web.mvc.ForwardResult\"/>\x0A</f></createCollections>\x0A\x09<mongodb public=\"1\" set=\"method\" line=\"121\"><f a=\"auth\">\x0A\x09<c path=\"String\"/>\x0A\x09<c path=\"ufront.web.mvc.ContentResult\"/>\x0A</f></mongodb>\x0A\x09<topRenderables public=\"1\" set=\"method\" line=\"170\"><f a=\"auth:?top\">\x0A\x09<c path=\"String\"/>\x0A\x09<c path=\"Int\"/>\x0A\x09<c path=\"ufront.web.mvc.ContentResult\"/>\x0A</f></topRenderables>\x0A\x09<purge public=\"1\" set=\"method\" line=\"184\"><f a=\"auth\">\x0A\x09<c path=\"String\"/>\x0A\x09<c path=\"ufront.web.mvc.ForwardResult\"/>\x0A</f></purge>\x0A\x09<purgeCache public=\"1\" set=\"method\" line=\"194\"><f a=\"auth\">\x0A\x09<c path=\"String\"/>\x0A\x09<c path=\"ufront.web.mvc.ForwardResult\"/>\x0A</f></purgeCache>\x0A\x09<clearCache public=\"1\" set=\"method\" line=\"202\"><f a=\"auth\">\x0A\x09<c path=\"String\"/>\x0A\x09<c path=\"ufront.web.mvc.ForwardResult\"/>\x0A</f></clearCache>\x0A\x09<purgeRenderables public=\"1\" set=\"method\" line=\"210\"><f a=\"auth\">\x0A\x09<c path=\"String\"/>\x0A\x09<c path=\"ufront.web.mvc.ForwardResult\"/>\x0A</f></purgeRenderables>\x0A\x09<purgeExpiredRenderables public=\"1\" set=\"method\" line=\"218\"><f a=\"auth\">\x0A\x09<c path=\"String\"/>\x0A\x09<c path=\"ufront.web.mvc.ForwardResult\"/>\x0A</f></purgeExpiredRenderables>\x0A\x09<info public=\"1\" set=\"method\" line=\"226\"><f a=\"\"><c path=\"String\"/></f></info>\x0A\x09<collectPhpInfo set=\"method\" line=\"231\"><f a=\"\"><c path=\"String\"/></f></collectPhpInfo>\x0A\x09<new public=\"1\" set=\"method\" line=\"14\"><f a=\"mongo\">\x0A\x09<c path=\"mongo.Mongo\"/>\x0A\x09<e path=\"Void\"/>\x0A</f></new>\x0A</class>";
 	function __toString() { return 'controller.SetupController'; }
 }
